backend/requirements.txt
fastapi
langchain
langchain-openai
uvicorn
websockets
openai
sounddevice
scipy
Pillow
qrcode[pil]
python-dotenv
numpy
torch
openai-whisper 